+++ mne/decoding/__init__.py
@@ -1,5 +1,6 @@
 from .transformer import Scaler, FilterEstimator
 from .transformer import PSDEstimator, ConcatenateChannels
+from .transformer import ConcatenateChannels as EpochVectorizer
 from .mixin import TransformerMixin
 from .base import BaseEstimator, LinearModel
 from .csp import CSP